+++ src/Config.php
@@ -170,16 +170,25 @@ class Config
         $password = Utils::replacePlaceholdersPassword($accountCfg['password'] ?? '');
 
         if ($password == "%b") {
-            if (
-                isset($_SESSION['oauth_token'])
-                && is_array($_SESSION['oauth_token'])
-                && isset($_SESSION['oauth_token']['access_token'])
-                && is_string($_SESSION['oauth_token']['access_token'])
-            ) {
-                $httpOptions = [ 'bearertoken' => $_SESSION['oauth_token']['access_token'] ];
-            } else {
+            $acctoken = false;
+            if (isset($_SESSION['oauth_token']) && is_array($_SESSION['oauth_token'])) {
+                $rcube = rcube::get_instance();
+                $acctoken = $rcube->decrypt((string) $_SESSION['password']);
+
+                if ($acctoken !== false) {
+                    // authorization stored in SESSION[password] is like "Bearer <token>", so cut the token type
+                    $token_type = ((string) $_SESSION['oauth_token']['token_type']) . " ";
+                    $token_type_len = strlen($token_type);
+                    if (strncmp($acctoken, $token_type, $token_type_len) === 0) {
+                        $acctoken = substr($acctoken, $token_type_len);
+                    }
+                }
+            }
+
+            if ($acctoken === false) {
                 throw new Exception("OAUTH2 bearer authentication requested, but no token available in roundcube");
             }
+            $httpOptions = [ 'bearertoken' => $acctoken ];
         } else {
             $username = Utils::replacePlaceholdersUsername($accountCfg['username'] ?? '');
             $httpOptions = [ 'username' => $username, 'password' => $password ];